chore(cubeorchestrator): split parser benches into their own target
